requirements.txt
# this is a backend requirments need when run
pip install flask-jwt-extended
python-dotenv
Flask
Flask-SQLAlchemy
SQLAlchemy
PyMySQL
flask-cors